Move position attention to ABIVisionDecoder
