+++ mmocr/models/common/modules/transformer_module.py
@@ -2,7 +2,6 @@ import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from mmcv.cnn import ConvModule
 
 from mmocr.models.builder import build_activation_layer
 
@@ -162,132 +161,3 @@ class PositionalEncoding(nn.Module):
         self.device = x.device
         x = x + self.position_table[:, :x.size(1)].clone().detach()
         return self.dropout(x)
-
-
-class PositionAttention(nn.Module):
-    """Transcribes visual features into character probabilities. A part of
-    implementation of `ABINet <https://arxiv.org/abs/1910.04396>`_. Adapted
-    from https://github.com/FangShancheng/ABINet .
-
-    Args:
-        max_length (int): Maximum length :math:`T` of generated text sequence.
-        in_channels (int): Number of channels :math:`E` of input vector.
-        num_channels (int): Number of channels of hidden vectors in mini U-Net.
-        h (int): Height :math:`H` of input image features.
-        w (int): Width :math:`W` of input image features.
-        mode (str): Upsampling mode for :obj:`torch.nn.Upsample` in mini U-Net.
-            Defaults to "nearest".
-    """
-
-    def __init__(self,
-                 max_length,
-                 in_channels=512,
-                 num_channels=64,
-                 h=8,
-                 w=32,
-                 mode='nearest',
-                 **kwargs):
-        super().__init__()
-        self.max_length = max_length
-
-        self.k_encoder = nn.Sequential(
-            self._encoder_layer(in_channels, num_channels, stride=(1, 2)),
-            self._encoder_layer(num_channels, num_channels, stride=(2, 2)),
-            self._encoder_layer(num_channels, num_channels, stride=(2, 2)),
-            self._encoder_layer(num_channels, num_channels, stride=(2, 2)))
-
-        self.k_decoder = nn.Sequential(
-            self._decoder_layer(
-                num_channels, num_channels, scale_factor=2, mode=mode),
-            self._decoder_layer(
-                num_channels, num_channels, scale_factor=2, mode=mode),
-            self._decoder_layer(
-                num_channels, num_channels, scale_factor=2, mode=mode),
-            self._decoder_layer(
-                num_channels, in_channels, size=(h, w), mode=mode))
-
-        self.pos_encoder = PositionalEncoding(in_channels, max_length)
-        self.project = nn.Linear(in_channels, in_channels)
-
-    def forward(self, x):
-        r"""
-        Args:
-            x (Tensor): Image features of shape (N, E, H, W).
-
-        Returns:
-            A tuple(attn_vecs, attn_scores).
-            attn_scores: Tensor of shape (N, T, H, W). Computed by
-                :math:`softmax(\frac{QK^T}{\sqrt{E}})` where :math:`Q` are
-                position encodings, :math:`K` are the features extracted from
-                input by mini U-net and :math:`E` is the number of channels.
-            attn_vecs: Tensor of shape (N, T, E). Computed by
-                :math:`attn_scores \cdot V` where :math:`V` are the same as
-                input features.
-        """
-        N, E, H, W = x.size()
-        k, v = x, x  # (N, E, H, W)
-
-        # Apply mini U-Net on k
-        features = []
-        for i in range(len(self.k_encoder)):
-            k = self.k_encoder[i](k)
-            features.append(k)
-        for i in range(len(self.k_decoder) - 1):
-            k = self.k_decoder[i](k)
-            k = k + features[len(self.k_decoder) - 2 - i]
-        k = self.k_decoder[-1](k)
-
-        # q = positional encoding
-        zeros = x.new_zeros((N, self.max_length, E))  # (N, T, E)
-        q = self.pos_encoder(zeros)  # (N, T, E)
-        q = self.project(q)  # (N, T, E)
-
-        # Attention encoding
-        attn_scores = torch.bmm(q, k.flatten(2, 3))  # (N, T, (H*W))
-        attn_scores = attn_scores / (E**0.5)
-        attn_scores = torch.softmax(attn_scores, dim=-1)
-
-        v = v.permute(0, 2, 3, 1).view(N, -1, E)  # (N, (H*W), E)
-        attn_vecs = torch.bmm(attn_scores, v)  # (N, T, E)
-
-        return attn_vecs, attn_scores.view(N, -1, H, W)
-
-    def _encoder_layer(self,
-                       in_channels,
-                       out_channels,
-                       kernel_size=3,
-                       stride=2,
-                       padding=1):
-        return ConvModule(
-            in_channels,
-            out_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding,
-            norm_cfg=dict(type='BN'),
-            act_cfg=dict(type='ReLU'))
-
-    def _decoder_layer(self,
-                       in_channels,
-                       out_channels,
-                       kernel_size=3,
-                       stride=1,
-                       padding=1,
-                       mode='nearest',
-                       scale_factor=None,
-                       size=None):
-        align_corners = None if mode == 'nearest' else True
-        return nn.Sequential(
-            nn.Upsample(
-                size=size,
-                scale_factor=scale_factor,
-                mode=mode,
-                align_corners=align_corners),
-            ConvModule(
-                in_channels,
-                out_channels,
-                kernel_size=kernel_size,
-                stride=stride,
-                padding=padding,
-                norm_cfg=dict(type='BN'),
-                act_cfg=dict(type='ReLU')))
