+++ mmocr/models/textrecog/decoders/abinet_vision_decoder.py
@@ -1,8 +1,10 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+import torch
 import torch.nn as nn
+from mmcv.cnn import ConvModule
 
 from mmocr.models.builder import DECODERS
-from mmocr.models.common.modules import PositionAttention
+from mmocr.models.common.modules import PositionalEncoding
 from .base_decoder import BaseDecoder
 
 
@@ -42,14 +44,30 @@ class ABIVisionDecoder(BaseDecoder):
                  **kwargs):
         super().__init__(init_cfg=init_cfg)
 
-        self.attention = PositionAttention(
-            max_length=max_seq_len,
-            in_channels=in_channels,
-            num_channels=num_channels,
-            mode=attn_mode,
-            h=attn_height,
-            w=attn_width,
-        )
+        self.max_seq_len = max_seq_len
+
+        # For mini-Unet
+        self.k_encoder = nn.Sequential(
+            self._encoder_layer(in_channels, num_channels, stride=(1, 2)),
+            self._encoder_layer(num_channels, num_channels, stride=(2, 2)),
+            self._encoder_layer(num_channels, num_channels, stride=(2, 2)),
+            self._encoder_layer(num_channels, num_channels, stride=(2, 2)))
+
+        self.k_decoder = nn.Sequential(
+            self._decoder_layer(
+                num_channels, num_channels, scale_factor=2, mode=attn_mode),
+            self._decoder_layer(
+                num_channels, num_channels, scale_factor=2, mode=attn_mode),
+            self._decoder_layer(
+                num_channels, num_channels, scale_factor=2, mode=attn_mode),
+            self._decoder_layer(
+                num_channels,
+                in_channels,
+                size=(attn_height, attn_width),
+                mode=attn_mode))
+
+        self.pos_encoder = PositionalEncoding(in_channels, max_seq_len)
+        self.project = nn.Linear(in_channels, in_channels)
         self.cls = nn.Linear(in_channels, num_chars)
 
     def forward_train(self,
@@ -71,14 +89,79 @@ class ABIVisionDecoder(BaseDecoder):
             - | attn_scores (Tensor): Shape (N, T, H, W). Intermediate result
                 for vision-language aligner.
         """
-        attn_vecs, attn_scores = self.attention(feat)
+        # Position Attention
+        N, E, H, W = feat.size()
+        k, v = feat, feat  # (N, E, H, W)
+
+        # Apply mini U-Net on k
+        features = []
+        for i in range(len(self.k_encoder)):
+            k = self.k_encoder[i](k)
+            features.append(k)
+        for i in range(len(self.k_decoder) - 1):
+            k = self.k_decoder[i](k)
+            k = k + features[len(self.k_decoder) - 2 - i]
+        k = self.k_decoder[-1](k)
+
+        # q = positional encoding
+        zeros = feat.new_zeros((N, self.max_seq_len, E))  # (N, T, E)
+        q = self.pos_encoder(zeros)  # (N, T, E)
+        q = self.project(q)  # (N, T, E)
+
+        # Attention encoding
+        attn_scores = torch.bmm(q, k.flatten(2, 3))  # (N, T, (H*W))
+        attn_scores = attn_scores / (E**0.5)
+        attn_scores = torch.softmax(attn_scores, dim=-1)
+        v = v.permute(0, 2, 3, 1).view(N, -1, E)  # (N, (H*W), E)
+        attn_vecs = torch.bmm(attn_scores, v)  # (N, T, E)
+
         logits = self.cls(attn_vecs)
         result = {
             'feature': attn_vecs,
             'logits': logits,
-            'attn_scores': attn_scores
+            'attn_scores': attn_scores.view(N, -1, H, W)
         }
         return result
 
     def forward_test(self, feat, out_enc=None, img_metas=None):
         return self.forward_train(feat, out_enc=out_enc, img_metas=img_metas)
+
+    def _encoder_layer(self,
+                       in_channels,
+                       out_channels,
+                       kernel_size=3,
+                       stride=2,
+                       padding=1):
+        return ConvModule(
+            in_channels,
+            out_channels,
+            kernel_size=kernel_size,
+            stride=stride,
+            padding=padding,
+            norm_cfg=dict(type='BN'),
+            act_cfg=dict(type='ReLU'))
+
+    def _decoder_layer(self,
+                       in_channels,
+                       out_channels,
+                       kernel_size=3,
+                       stride=1,
+                       padding=1,
+                       mode='nearest',
+                       scale_factor=None,
+                       size=None):
+        align_corners = None if mode == 'nearest' else True
+        return nn.Sequential(
+            nn.Upsample(
+                size=size,
+                scale_factor=scale_factor,
+                mode=mode,
+                align_corners=align_corners),
+            ConvModule(
+                in_channels,
+                out_channels,
+                kernel_size=kernel_size,
+                stride=stride,
+                padding=padding,
+                norm_cfg=dict(type='BN'),
+                act_cfg=dict(type='ReLU')))
