+++ tests/test_models/test_ocr_layer.py
@@ -3,7 +3,6 @@ import torch
 
 from mmocr.models.common import (PositionalEncoding, TFDecoderLayer,
                                  TFEncoderLayer)
-from mmocr.models.common.modules import PositionAttention
 from mmocr.models.textrecog.layers import BasicBlock, Bottleneck
 from mmocr.models.textrecog.layers.conv_layer import conv3x3
 
@@ -62,11 +61,3 @@ def test_transformer_layer():
         operation_order=('self_attn', 'norm', 'ffn', 'norm'))
     out_enc = encoder_layer(in_enc)
     assert out_dec.shape == torch.Size([1, 30, 512])
-
-
-def test_position_attention():
-    layer = PositionAttention(32, 512, 64, 8, 32)
-    x = torch.randn(2, 512, 8, 32)
-    attn_vecs, attn_scores = layer(x)
-    assert attn_vecs.shape == torch.Size([2, 32, 512])
-    assert attn_scores.shape == torch.Size([2, 32, 8, 32])
